+++ pkg/mineral/mineral.go
@@ -3,6 +3,7 @@ package mineral
 import (
 	"math/rand"
 
+	"github.com/ironarachne/world/pkg/random"
 	"github.com/ironarachne/world/pkg/resource"
 )
 
@@ -12,6 +13,7 @@ type Mineral struct {
 	PluralName   string
 	Hardness     int
 	Malleability int
+	Commonality  int
 	Resources    []resource.Resource
 }
 
@@ -26,9 +28,45 @@ func Random(amount int, from []Mineral) []Mineral {
 	}
 
 	for i := 0; i < amount; i++ {
-		mineral = from[rand.Intn(len(from)-1)]
+		mineral = from[rand.Intn(len(from))]
 		if !InSlice(mineral, minerals) {
 			minerals = append(minerals, mineral)
+		}
+	}
+
+	return minerals
+}
+
+// RandomWeighted returns a random mineral from a slice, considering weights
+func RandomWeighted(from []Mineral) Mineral {
+	names := make(map[string]int)
+
+	for _, m := range from {
+		names[m.Name] = m.Commonality
+	}
+
+	name := random.StringFromThresholdMap(names)
+
+	for _, m := range from {
+		if m.Name == name {
+			return m
+		}
+	}
+
+	panic("Couldn't find named mineral")
+}
+
+// RandomWeightedSet returns a set of random weighted minerals
+func RandomWeightedSet(amount int, from []Mineral) []Mineral {
+	var minerals []Mineral
+	var mineral Mineral
+
+	for i := 0; i < amount; i++ {
+		mineral = RandomWeighted(from)
+		if !InSlice(mineral, minerals) {
+			minerals = append(minerals, mineral)
+		} else {
+			i--
 		}
 	}