+++ pkg/resource/jewelry.go
@@ -21,6 +21,76 @@ func getJewelry() []Pattern {
 				},
 			},
 		},
+		{
+			Name:        "necklace",
+			Description: "a necklace",
+			Tags: []string{
+				"necklace",
+			},
+			Profession: producer,
+			Slots: []Slot{
+				{
+					Name:                "body",
+					RequiredTag:         "metal ingot",
+					DescriptionTemplate: "{{.Resource.Origin}} necklace",
+				},
+			},
+		},
+		{
+			Name:        "pendant",
+			Description: "a pendant",
+			Tags: []string{
+				"pendant",
+			},
+			Profession: producer,
+			Slots: []Slot{
+				{
+					Name:                "gem",
+					RequiredTag:         "gem",
+					DescriptionTemplate: "{{.Resource.Origin}} pendant",
+				},
+				{
+					Name:                "thong",
+					RequiredTag:         "leather",
+					DescriptionTemplate: " on a {{.Resource.Origin}} leather thong",
+				},
+			},
+		},
+		{
+			Name:        "ring",
+			Description: "a ring",
+			Tags: []string{
+				"ring",
+			},
+			Profession: producer,
+			Slots: []Slot{
+				{
+					Name:                "band",
+					RequiredTag:         "metal ingot",
+					DescriptionTemplate: "{{.Resource.Origin}} ring",
+				},
+				{
+					Name:                "gem",
+					RequiredTag:         "gem",
+					DescriptionTemplate: " set with {{.Resource.Origin}}",
+				},
+			},
+		},
+		{
+			Name:        "ring",
+			Description: "a ring",
+			Tags: []string{
+				"ring",
+			},
+			Profession: producer,
+			Slots: []Slot{
+				{
+					Name:                "band",
+					RequiredTag:         "metal ingot",
+					DescriptionTemplate: "{{.Resource.Origin}} ring",
+				},
+			},
+		},
 	}
 
 	return patterns