+++ pkg/resource/logging.go
@@ -9,6 +9,21 @@ func getLogging() []Pattern {
 
 	patterns := []Pattern{
 		{
+			Name:        "wood log",
+			Description: "a trimmed wood log",
+			Tags: []string{
+				"wood log",
+			},
+			Profession: producer,
+			Slots: []Slot{
+				{
+					Name:                "body",
+					RequiredTag:         "wood",
+					DescriptionTemplate: "{{.Resource.Origin}} log",
+				},
+			},
+		},
+		{
 			Name:        "wood plank",
 			Description: "a rectangular length of wood",
 			Tags: []string{